+++ Source/Scene/Scene.js
@@ -51,9 +51,11 @@ define([
         '../Renderer/ShaderProgram',
         '../Renderer/ShaderSource',
         '../Renderer/Texture',
+        '../ThirdParty/when',
         './BrdfLutGenerator',
         './Camera',
         './Cesium3DTileFeature',
+        './Cesium3DTileset',
         './CreditDisplay',
         './DebugCameraPrimitive',
         './DepthPlane',
@@ -132,9 +134,11 @@ define([
         ShaderProgram,
         ShaderSource,
         Texture,
+        when,
         BrdfLutGenerator,
         Camera,
         Cesium3DTileFeature,
+        Cesium3DTileset,
         CreditDisplay,
         DebugCameraPrimitive,
         DepthPlane,
@@ -169,6 +173,14 @@ define([
             });
         };
     };
+
+    function AsyncRayPick(ray, primitives) {
+        this.ray = ray;
+        this.primitives = primitives;
+        this.ready = false;
+        this.deferred = when.defer();
+        this.promise = this.deferred.promise;
+    }
 
     /**
      * The container for all 3D graphical objects and state in a Cesium virtual scene.  Generally,
@@ -282,6 +294,8 @@ define([
         this._globe = undefined;
         this._primitives = new PrimitiveCollection();
         this._groundPrimitives = new PrimitiveCollection();
+
+        this._asyncRayPicks = [];
 
         this._logDepthBuffer = context.fragmentDepth;
         this._logDepthBufferDirty = true;
@@ -892,13 +906,14 @@ define([
         },
 
         /**
-         * Returns <code>true</code> if the {@link Scene#sampleHeight} function is supported.
+         * Returns <code>true</code> if the {@link Scene#sampleHeight} and {@link Scene#sampleHeightMostDetailed} functions are supported.
          * @memberof Scene.prototype
          *
          * @type {Boolean}
          * @readonly
          *
          * @see Scene#sampleHeight
+         * @see Scene#sampleHeightMostDetailed
          */
         sampleHeightSupported : {
             get : function() {
@@ -907,15 +922,31 @@ define([
         },
 
         /**
-         * Returns <code>true</code> if the {@link Scene#clampToHeight} function is supported.
+         * Returns <code>true</code> if the {@link Scene#clampToHeight} and {@link Scene#clampToHeightMostDetailed} functions are supported.
          * @memberof Scene.prototype
          *
          * @type {Boolean}
          * @readonly
          *
          * @see Scene#clampToHeight
+         * @see Scene#clampToHeightMostDetailed
          */
         clampToHeightSupported : {
+            get : function() {
+                return this._context.depthTexture;
+            }
+        },
+
+        /**
+         * Returns <code>true</code> if the {@link Scene#invertClassification} is supported.
+         * @memberof Scene.prototype
+         *
+         * @type {Boolean}
+         * @readonly
+         *
+         * @see Scene#invertClassification
+         */
+        invertClassificationSupported : {
             get : function() {
                 return this._context.depthTexture;
             }
@@ -1578,6 +1609,7 @@ define([
         passes.depth = false;
         passes.postProcess = false;
         passes.offscreen = false;
+        passes.async = false;
     }
 
     function updateFrameNumber(scene, frameNumber, time) {
@@ -3009,6 +3041,8 @@ define([
             scene.globe.update(frameState);
         }
 
+        updateAsyncRayPicks(scene);
+
         frameState.creditDisplay.update();
     }
 
@@ -3650,6 +3684,85 @@ define([
         camera.right = right;
     }
 
+    function updateAsyncRayPick(scene, asyncRayPick) {
+        var context = scene._context;
+        var uniformState = context.uniformState;
+        var frameState = scene._frameState;
+
+        var view = scene._pickOffscreenView;
+        scene._view = view;
+
+        var ray = asyncRayPick.ray;
+        var primitives = asyncRayPick.primitives;
+
+        updateCameraFromRay(ray, view.camera);
+
+        updateFrameState(scene);
+        frameState.passes.offscreen = true;
+        frameState.passes.async = true;
+
+        uniformState.update(frameState);
+
+        var commandList = frameState.commandList;
+        var commandsLength = commandList.length;
+
+        var ready = true;
+        var primitivesLength = primitives.length;
+        for (var i = 0; i < primitivesLength; ++i) {
+            var primitive = primitives[i];
+            if (primitive.show && scene.primitives.contains(primitive)) {
+                // Only update primitives that are still contained in the scene's primitive collection and are still visible
+                // Update primitives continually until all primitives are ready. This way tiles are never removed from the cache.
+                var primitiveReady = primitive.updateAsync(frameState);
+                ready = (ready && primitiveReady);
+            }
+        }
+
+        // Ignore commands pushed during async pass
+        commandList.length = commandsLength;
+
+        scene._view = scene._defaultView;
+
+        if (ready) {
+            asyncRayPick.deferred.resolve();
+        }
+
+        return ready;
+    }
+
+    function updateAsyncRayPicks(scene) {
+        // Modifies array during iteration
+        var asyncRayPicks = scene._asyncRayPicks;
+        for (var i = 0; i < asyncRayPicks.length; ++i) {
+            if (updateAsyncRayPick(scene, asyncRayPicks[i])) {
+                asyncRayPicks.splice(i--, 1);
+            }
+        }
+    }
+
+    function launchAsyncRayPick(scene, ray, objectsToExclude, callback) {
+        var asyncPrimitives = [];
+        var primitives = scene.primitives;
+        var length = primitives.length;
+        for (var i = 0; i < length; ++i) {
+            var primitive = primitives.get(i);
+            if ((primitive instanceof Cesium3DTileset) && primitive.show) {
+                if (!defined(objectsToExclude) || objectsToExclude.indexOf(primitive) === -1) {
+                    asyncPrimitives.push(primitive);
+                }
+            }
+        }
+        if (asyncPrimitives.length === 0) {
+            return when.resolve(callback());
+        }
+
+        var asyncRayPick = new AsyncRayPick(ray, asyncPrimitives);
+        scene._asyncRayPicks.push(asyncRayPick);
+        return asyncRayPick.promise.then(function() {
+            return callback();
+        });
+    }
+
     function isExcluded(object, objectsToExclude) {
         if (!defined(object) || !defined(objectsToExclude) || objectsToExclude.length === 0) {
             return false;
@@ -3659,7 +3772,7 @@ define([
                (objectsToExclude.indexOf(object.id) > -1);
     }
 
-    function getRayIntersection(scene, ray, objectsToExclude, requirePosition) {
+    function getRayIntersection(scene, ray, objectsToExclude, requirePosition, async) {
         var context = scene._context;
         var uniformState = context.uniformState;
         var frameState = scene._frameState;
@@ -3679,6 +3792,7 @@ define([
         frameState.invertClassification = false;
         frameState.passes.pick = true;
         frameState.passes.offscreen = true;
+        frameState.passes.async = async;
 
         uniformState.update(frameState);
 
@@ -3717,22 +3831,22 @@ define([
         }
     }
 
-    function getRayIntersections(scene, ray, limit, objectsToExclude, requirePosition) {
+    function getRayIntersections(scene, ray, limit, objectsToExclude, requirePosition, async) {
         var pickCallback = function() {
-            return getRayIntersection(scene, ray, objectsToExclude, requirePosition);
+            return getRayIntersection(scene, ray, objectsToExclude, requirePosition, async);
         };
         return drillPick(limit, pickCallback);
     }
 
-    function pickFromRay(scene, ray, objectsToExclude, requirePosition) {
-        var results = getRayIntersections(scene, ray, 1, objectsToExclude, requirePosition);
+    function pickFromRay(scene, ray, objectsToExclude, requirePosition, async) {
+        var results = getRayIntersections(scene, ray, 1, objectsToExclude, requirePosition, async);
         if (results.length > 0) {
             return results[0];
         }
     }
 
-    function drillPickFromRay(scene, ray, limit, objectsToExclude, requirePosition) {
-        return getRayIntersections(scene, ray, limit, objectsToExclude, requirePosition);
+    function drillPickFromRay(scene, ray, limit, objectsToExclude, requirePosition, async) {
+        return getRayIntersections(scene, ray, limit, objectsToExclude, requirePosition, async);
     }
 
     /**
@@ -3748,7 +3862,7 @@ define([
      * @private
      *
      * @param {Ray} ray The ray.
-     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or features to exclude from the ray intersection.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to exclude from the ray intersection.
      * @returns {Object} An object containing the object and position of the first intersection.
      *
      * @exception {DeveloperError} Ray intersections are only supported in 3D mode.
@@ -3760,7 +3874,7 @@ define([
             throw new DeveloperError('Ray intersections are only supported in 3D mode.');
         }
         //>>includeEnd('debug');
-        return pickFromRay(this, ray, objectsToExclude, false);
+        return pickFromRay(this, ray, objectsToExclude, false, false);
     };
 
     /**
@@ -3778,7 +3892,7 @@ define([
      *
      * @param {Ray} ray The ray.
      * @param {Number} [limit=Number.MAX_VALUE] If supplied, stop finding intersections after this many intersections.
-     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or features to exclude from the ray intersection.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to exclude from the ray intersection.
      * @returns {Object[]} List of objects containing the object and position of each intersection.
      *
      * @exception {DeveloperError} Ray intersections are only supported in 3D mode.
@@ -3790,7 +3904,62 @@ define([
             throw new DeveloperError('Ray intersections are only supported in 3D mode.');
         }
         //>>includeEnd('debug');
-        return drillPickFromRay(this, ray, limit, objectsToExclude, false);
+        return drillPickFromRay(this, ray, limit, objectsToExclude, false, false);
+    };
+
+    /**
+     * Initiates an asynchronous {@link Scene#pickFromRay} request using the maximum level of detail for 3D Tilesets
+     * regardless of visibility.
+     *
+     * @private
+     *
+     * @param {Ray} ray The ray.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to exclude from the ray intersection.
+     * @returns {Promise.<Object>} A promise that resolves to an object containing the object and position of the first intersection.
+     *
+     * @exception {DeveloperError} Ray intersections are only supported in 3D mode.
+     */
+    Scene.prototype.pickFromRayMostDetailed = function(ray, objectsToExclude) {
+        //>>includeStart('debug', pragmas.debug);
+        Check.defined('ray', ray);
+        if (this._mode !== SceneMode.SCENE3D) {
+            throw new DeveloperError('Ray intersections are only supported in 3D mode.');
+        }
+        //>>includeEnd('debug');
+        var that = this;
+        ray = Ray.clone(ray);
+        objectsToExclude = defined(objectsToExclude) ? objectsToExclude.slice() : objectsToExclude;
+        return launchAsyncRayPick(this, ray, objectsToExclude, function() {
+            return pickFromRay(that, ray, objectsToExclude, false, true);
+        });
+    };
+
+    /**
+     * Initiates an asynchronous {@link Scene#drillPickFromRay} request using the maximum level of detail for 3D Tilesets
+     * regardless of visibility.
+     *
+     * @private
+     *
+     * @param {Ray} ray The ray.
+     * @param {Number} [limit=Number.MAX_VALUE] If supplied, stop finding intersections after this many intersections.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to exclude from the ray intersection.
+     * @returns {Promise.<Object[]>} A promise that resolves to a list of objects containing the object and position of each intersection.
+     *
+     * @exception {DeveloperError} Ray intersections are only supported in 3D mode.
+     */
+    Scene.prototype.drillPickFromRayMostDetailed = function(ray, limit, objectsToExclude) {
+        //>>includeStart('debug', pragmas.debug);
+        Check.defined('ray', ray);
+        if (this._mode !== SceneMode.SCENE3D) {
+            throw new DeveloperError('Ray intersections are only supported in 3D mode.');
+        }
+        //>>includeEnd('debug');
+        var that = this;
+        ray = Ray.clone(ray);
+        objectsToExclude = defined(objectsToExclude) ? objectsToExclude.slice() : objectsToExclude;
+        return launchAsyncRayPick(this, ray, objectsToExclude, function() {
+            return drillPickFromRay(that, ray, limit, objectsToExclude, false, true);
+        });
     };
 
     var scratchSurfacePosition = new Cartesian3();
@@ -3827,19 +3996,47 @@ define([
         return cartographic.height;
     }
 
+    function sampleHeightMostDetailed(scene, cartographic, objectsToExclude) {
+        var ray = getRayForSampleHeight(scene, cartographic);
+        return launchAsyncRayPick(scene, ray, objectsToExclude, function() {
+            var pickResult = pickFromRay(scene, ray, objectsToExclude, true, true);
+            if (defined(pickResult)) {
+                return getHeightFromCartesian(scene, pickResult.position);
+            }
+        });
+    }
+
+    function clampToHeightMostDetailed(scene, cartesian, objectsToExclude, result) {
+        var ray = getRayForClampToHeight(scene, cartesian);
+        return launchAsyncRayPick(scene, ray, objectsToExclude, function() {
+            var pickResult = pickFromRay(scene, ray, objectsToExclude, true, true);
+            if (defined(pickResult)) {
+                return Cartesian3.clone(pickResult.position, result);
+            }
+        });
+    }
+
     /**
      * Returns the height of scene geometry at the given cartographic position or <code>undefined</code> if there was no
-     * scene geometry to sample height from. May be used to clamp objects to the globe, 3D Tiles, or primitives in the scene.
+     * scene geometry to sample height from. The height of the input position is ignored. May be used to clamp objects to
+     * the globe, 3D Tiles, or primitives in the scene.
      * <p>
      * This function only samples height from globe tiles and 3D Tiles that are rendered in the current view. Samples height
      * from all other primitives regardless of their visibility.
      * </p>
      *
      * @param {Cartographic} position The cartographic position to sample height from.
-     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or features to not sample height from.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to not sample height from.
      * @returns {Number} The height. This may be <code>undefined</code> if there was no scene geometry to sample height from.
      *
+     * @example
+     * var position = new Cesium.Cartographic(-1.31968, 0.698874);
+     * var height = viewer.scene.sampleHeight(position);
+     * console.log(height);
+     *
      * @see Scene#clampToHeight
+     * @see Scene#clampToHeightMostDetailed
+     * @see Scene#sampleHeightMostDetailed
      *
      * @exception {DeveloperError} sampleHeight is only supported in 3D mode.
      * @exception {DeveloperError} sampleHeight requires depth texture support. Check sampleHeightSupported.
@@ -3855,7 +4052,7 @@ define([
         }
         //>>includeEnd('debug');
         var ray = getRayForSampleHeight(this, position);
-        var pickResult = pickFromRay(this, ray, objectsToExclude, true);
+        var pickResult = pickFromRay(this, ray, objectsToExclude, true, false);
         if (defined(pickResult)) {
             return getHeightFromCartesian(this, pickResult.position);
         }
@@ -3871,11 +4068,18 @@ define([
      * </p>
      *
      * @param {Cartesian3} cartesian The cartesian position.
-     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or features to not clamp to.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to not clamp to.
      * @param {Cartesian3} [result] An optional object to return the clamped position.
      * @returns {Cartesian3} The modified result parameter or a new Cartesian3 instance if one was not provided. This may be <code>undefined</code> if there was no scene geometry to clamp to.
      *
+     * @example
+     * // Clamp an entity to the underlying scene geometry
+     * var position = entity.position.getValue(Cesium.JulianDate.now());
+     * entity.position = viewer.scene.clampToHeight(position);
+     *
      * @see Scene#sampleHeight
+     * @see Scene#sampleHeightMostDetailed
+     * @see Scene#clampToHeightMostDetailed
      *
      * @exception {DeveloperError} clampToHeight is only supported in 3D mode.
      * @exception {DeveloperError} clampToHeight requires depth texture support. Check clampToHeightSupported.
@@ -3891,10 +4095,113 @@ define([
         }
         //>>includeEnd('debug');
         var ray = getRayForClampToHeight(this, cartesian);
-        var pickResult = pickFromRay(this, ray, objectsToExclude, true);
+        var pickResult = pickFromRay(this, ray, objectsToExclude, true, false);
         if (defined(pickResult)) {
             return Cartesian3.clone(pickResult.position, result);
         }
+    };
+
+    /**
+     * Initiates an asynchronous {@link Scene#sampleHeight} query for an array of {@link Cartographic} positions
+     * using the maximum level of detail for 3D Tilesets in the scene. The height of the input positions is ignored.
+     * Returns a promise that is resolved when the query completes. Each point height is modified in place.
+     * If a height cannot be determined because no geometry can be sampled at that location, or another error occurs,
+     * the height is set to undefined.
+     *
+     * @param {Cartographic[]} positions The cartographic positions to update with sampled heights.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to not sample height from.
+     * @returns {Promise.<Number[]>} A promise that resolves to the provided list of positions when the query has completed.
+     *
+     * @example
+     * var positions = [
+     *     new Cesium.Cartographic(-1.31968, 0.69887),
+     *     new Cesium.Cartographic(-1.10489, 0.83923)
+     * ];
+     * var promise = viewer.scene.sampleHeightMostDetailed(positions);
+     * promise.then(function(updatedPosition) {
+     *     // positions[0].height and positions[1].height have been updated.
+     *     // updatedPositions is just a reference to positions.
+     * }
+     *
+     * @see Scene#sampleHeight
+     *
+     * @exception {DeveloperError} sampleHeightMostDetailed is only supported in 3D mode.
+     * @exception {DeveloperError} sampleHeightMostDetailed requires depth texture support. Check sampleHeightSupported.
+     */
+    Scene.prototype.sampleHeightMostDetailed = function(positions, objectsToExclude) {
+        //>>includeStart('debug', pragmas.debug);
+        Check.defined('positions', positions);
+        if (this._mode !== SceneMode.SCENE3D) {
+            throw new DeveloperError('sampleHeightMostDetailed is only supported in 3D mode.');
+        }
+        if (!this.sampleHeightSupported) {
+            throw new DeveloperError('sampleHeightMostDetailed requires depth texture support. Check sampleHeightSupported.');
+        }
+        //>>includeEnd('debug');
+        objectsToExclude = defined(objectsToExclude) ? objectsToExclude.slice() : objectsToExclude;
+        var length = positions.length;
+        var promises = new Array(length);
+        for (var i = 0; i < length; ++i) {
+            promises[i] = sampleHeightMostDetailed(this, positions[i], objectsToExclude);
+        }
+        return when.all(promises).then(function(heights) {
+            var length = heights.length;
+            for (var i = 0; i < length; ++i) {
+                positions[i].height = heights[i];
+            }
+            return positions;
+        });
+    };
+
+    /**
+     * Initiates an asynchronous {@link Scene#clampToHeight} query for an array of {@link Cartesian3} positions
+     * using the maximum level of detail for 3D Tilesets in the scene. Returns a promise that is resolved when
+     * the query completes. Each position is modified in place. If a position cannot be clamped because no geometry
+     * can be sampled at that location, or another error occurs, the element in the array is set to undefined.
+     *
+     * @param {Cartesian3[]} cartesians The cartesian positions to update with clamped positions.
+     * @param {Object[]} [objectsToExclude] A list of primitives, entities, or 3D Tiles features to not clamp to.
+     * @returns {Promise.<Cartesian3[]>} A promise that resolves to the provided list of positions when the query has completed.
+     *
+     * @example
+     * var cartesians = [
+     *     entities[0].position.getValue(Cesium.JulianDate.now()),
+     *     entities[1].position.getValue(Cesium.JulianDate.now())
+     * ];
+     * var promise = viewer.scene.clampToHeightMostDetailed(cartesians);
+     * promise.then(function(updatedCartesians) {
+     *     entities[0].position = updatedCartesians[0];
+     *     entities[1].position = updatedCartesians[1];
+     * }
+     *
+     * @see Scene#clampToHeight
+     *
+     * @exception {DeveloperError} clampToHeightMostDetailed is only supported in 3D mode.
+     * @exception {DeveloperError} clampToHeightMostDetailed requires depth texture support. Check clampToHeightSupported.
+     */
+    Scene.prototype.clampToHeightMostDetailed = function(cartesians, objectsToExclude) {
+        //>>includeStart('debug', pragmas.debug);
+        Check.defined('cartesians', cartesians);
+        if (this._mode !== SceneMode.SCENE3D) {
+            throw new DeveloperError('clampToHeightMostDetailed is only supported in 3D mode.');
+        }
+        if (!this.clampToHeightSupported) {
+            throw new DeveloperError('clampToHeightMostDetailed requires depth texture support. Check clampToHeightSupported.');
+        }
+        //>>includeEnd('debug');
+        objectsToExclude = defined(objectsToExclude) ? objectsToExclude.slice() : objectsToExclude;
+        var length = cartesians.length;
+        var promises = new Array(length);
+        for (var i = 0; i < length; ++i) {
+            promises[i] = clampToHeightMostDetailed(this, cartesians[i], objectsToExclude, cartesians[i]);
+        }
+        return when.all(promises).then(function(clampedCartesians) {
+            var length = clampedCartesians.length;
+            for (var i = 0; i < length; ++i) {
+                cartesians[i] = clampedCartesians[i];
+            }
+            return cartesians;
+        });
     };
 
     /**